+++ internal/gateway/api/v1/application_handler_test.go
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package v1kubeflow
+package v1
 
 import (
 	"bytes"
@@ -112,14 +112,14 @@ func TestApplicationHandlerGet(t *testing.T) {
 		User:    "user",
 	}
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		GetFunc: func(ctx context.Context, gatewayId string) (*domain.GatewayApplication, error) {
 			return retApp, nil
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("GET", "/api/v1/applications/clusterid-testid", nil)
 	w := httptest.NewRecorder()
@@ -134,14 +134,14 @@ func TestApplicationHandlerGet(t *testing.T) {
 
 func TestApplicationHandlerGetError(t *testing.T) {
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		GetFunc: func(ctx context.Context, gatewayId string) (*domain.GatewayApplication, error) {
 			return &domain.GatewayApplication{}, gatewayerrors.NewNotFound(errors.New("error getting SparkApplication 'clusterid-testid'"))
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("GET", "/api/v1/applications/clusterid-testid", nil)
 	w := httptest.NewRecorder()
@@ -162,14 +162,14 @@ func TestApplicationHandlerStatus(t *testing.T) {
 		SubmissionID: "submissionId",
 	}
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		StatusFunc: func(ctx context.Context, gatewayId string) (*v1beta2.SparkApplicationStatus, error) {
 			return retResp, nil
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("GET", "/api/v1/applications/clusterid-testid/status", nil)
 	w := httptest.NewRecorder()
@@ -183,14 +183,14 @@ func TestApplicationHandlerStatus(t *testing.T) {
 }
 func TestApplicationHandlerStatusError(t *testing.T) {
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		StatusFunc: func(ctx context.Context, gatewayId string) (*v1beta2.SparkApplicationStatus, error) {
 			return &v1beta2.SparkApplicationStatus{}, gatewayerrors.NewNotFound(errors.New("error getting SparkApplication 'clusterid-testid'"))
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("GET", "/api/v1/applications/clusterid-testid/status", nil)
 	w := httptest.NewRecorder()
@@ -222,13 +222,13 @@ func TestApplicationHandlerCreate(t *testing.T) {
 		User:    "user",
 	}
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		CreateFunc: func(ctx context.Context, application *v1beta2.SparkApplication, user string) (*domain.GatewayApplication, error) {
 			return retApp, nil
 		},
 	}
 
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	createApp := &v1beta2.SparkApplication{
 		ObjectMeta: v1.ObjectMeta{
@@ -250,14 +250,14 @@ func TestApplicationHandlerCreate(t *testing.T) {
 }
 func TestApplicationHandlerCreateBadRequest(t *testing.T) {
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		CreateFunc: func(ctx context.Context, application *v1beta2.SparkApplication, user string) (*domain.GatewayApplication, error) {
 			return nil, nil
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("POST", "/api/v1/applications", nil)
 	w := httptest.NewRecorder()
@@ -278,13 +278,13 @@ func TestApplicationHandlerCreateAlreadyExists(t *testing.T) {
 		ctx.Next()
 	})
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		CreateFunc: func(ctx context.Context, application *v1beta2.SparkApplication, user string) (*domain.GatewayApplication, error) {
 			return nil, gatewayerrors.NewAlreadyExists(errors.New("resource.group \"test\" already exists"))
 		},
 	}
 
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	createReq := &v1beta2.SparkApplication{
 		ObjectMeta: v1.ObjectMeta{
@@ -307,14 +307,14 @@ func TestApplicationHandlerCreateAlreadyExists(t *testing.T) {
 
 func TestApplicationHandlerDelete(t *testing.T) {
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		DeleteFunc: func(ctx context.Context, gatewayId string) error {
 			return nil
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("DELETE", "/api/v1/applications/clusterid-testid", nil)
 	w := httptest.NewRecorder()
@@ -331,14 +331,14 @@ func TestApplicationHandlerDelete(t *testing.T) {
 }
 func TestApplicationHandlerDeleteError(t *testing.T) {
 
-	service := &service.SparkApplicationServiceMock{
+	service := &service.GatewayApplicationServiceMock{
 		DeleteFunc: func(ctx context.Context, gatewayId string) error {
 			return gatewayerrors.NewNotFound(errors.New("error getting SparkApplication 'clusterid-testid'"))
 		},
 	}
 
 	router, v1Group := NewV1Router()
-	RegisterKubeflowApplicationRoutes(v1Group, testConfig, service)
+	RegisterGatewayApplicationRoutes(v1Group, testConfig, service)
 
 	req, _ := http.NewRequest("DELETE", "/api/v1/applications/clusterid-testid", nil)
 	w := httptest.NewRecorder()
